07.02

+++ nucleo-target/Core/Inc/sw_timers.h
@@ -19,8 +19,8 @@
 struct timer 
 {
 	volatile bool set;
-	volatile uint32_t start;
-	volatile uint32_t interval;
+	volatile unsigned long long start;
+	volatile unsigned long long interval;
 	void (*callback)(void);
 	volatile bool on_interrupt;
 	volatile bool autorestart;
@@ -33,7 +33,7 @@ void Timer_restart(struct timer* t);
 void Timer_enable(struct timer* t);
 void Timer_disable(struct timer* t);
 bool Timer_expired(struct timer* t);
-uint32_t Timer_remaining(struct timer* t);
+unsigned long long Timer_remaining(struct timer* t);
 
 void t_OnDigitCompleteInterrupt(void);
 void t_OnDigitCompleteContinuous(void);
